refactor some gene selection related architecture and key management stuff
