+++ dynamo/preprocessing/__init__.py
@@ -13,7 +13,7 @@ from .preprocess import (
     estimate_dispersion,
     filter_cells,
     select_genes,
-    filter_genes,
+    filter_genes_by_outliers,
     filter_genes_by_clusters_,
     SVRs,
     get_svr_filter,
@@ -43,7 +43,7 @@ __all__ = [
     "estimate_dispersion",
     "filter_cells",
     "select_genes",
-    "filter_genes",
+    "filter_genes_by_outliers",
     "filter_genes_by_clusters_",
     "SVRs",
     "get_svr_filter",
